go mod: upgrade ginkgo to ginkgo/v2
The code.cloudfoundry.org/hwc module will now use ginkgo v2 for its
tests and the latest version of gomega.

Also updates GinkgoParallelNode to GinkgoParallelProcess because that
was deprecated in the upgrade to ginkgo v2.

Future note:
```
github.com/onsi/ginkgo v1.16.5 // indirect
```
This line is present in the go.mod file because
github.cloudfoundry-community/go-cfenv has not yet been upgraded to
ginkgo v2. Once that is updated the transient dependency on the old
ginkgo will be removed.